requirements.txt
openai>=1.0.0
aiohttp>=3.9.0
pandas>=2.0.0
numpy>=1.24.0
dnspython>=2.4.0
python-whois>=0.8.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
